+++ packages/editor/conversation/docs/0-intro.tsx
@@ -199,6 +199,7 @@ const props = {
               kind: 'generic',
               value: {
                 kind: 'function',
+                parameters: [],
                 returnType: {
                   kind: 'void',
                 },
@@ -223,6 +224,7 @@ const props = {
               kind: 'generic',
               value: {
                 kind: 'function',
+                parameters: [],
                 returnType: {
                   kind: 'generic',
                   value: {