+++ packages/editor/editor-mobile-bridge/src/web-to-native/bridge.ts
@@ -8,6 +8,17 @@ export interface TextFormattingBridge {
   updateText(content: string);
 }
 
+export interface MediaBridge {
+  getServiceHost(): string;
+  getCollection(): string;
+}
+
+export interface PromiseBridge {
+  submitPromise(name: string, uuid: string, args: string);
+}
+
 export default interface NativeBridge
   extends MentionBridge,
-    TextFormattingBridge {};
+    TextFormattingBridge,
+    MediaBridge,
+    PromiseBridge {};